pyproject.toml
[project]
name = "bingo-agent-simulation"
version = "0.1.0"
description = "A simulation of the bingo agent"
authors = [{ name = "Aniketh Vijesh" }, { name = "Sree Sharvesh"}]
dependencies = [
    "black>=25.1.0",
    "dotenv>=0.9.9",
    "hydra-core>=1.3.2",
    "langchain>=0.3.26",
    "langchain-google-genai>=0.1.0",
]
